+++ src/aim/config/__init__.py
@@ -1,4 +1,2 @@
-#from aim.config.config import Config
-from aim.config.cfg_s3 import S3Config
 from aim.config.cfg_codecommit import CodeCommitConfig
 from aim.config.cfg_route53 import Route53Config