+++ src/aim/stack_group/grp_application.py
@@ -36,6 +36,7 @@ class LaunchBundle():
                                          self.app_id,
                                          self.group_id,
                                          self.resource_id)
+        self.s3_bucket_ref = None
         self.s3_context_id = self.manager.subenv_ctx.gen_ref(
                                     app_id=self.app_id,
                                     grp_id=self.group_id,
@@ -111,8 +112,7 @@ class EC2LaunchManager():
     def stack_hook(self, hook, bundle):
         # Upload bundle to the S3 bucket
         s3_ctl = self.aim_ctx.get_controller('S3')
-        s3_ctl.init_context(bundle.s3_context_id)
-        bucket_name = s3_ctl.get_bucket_name(bundle.app_id, bundle.group_id, bundle.bucket_id)
+        bucket_name = s3_ctl.get_bucket_name(bundle.s3_context_id, bundle.s3_bucket_ref)
         s3_client = self.account_ctx.get_aws_client('s3')
         bundle_s3_key = os.path.join("LaunchBundles", bundle.package_filename)
         s3_client.upload_file(bundle.package_path, bucket_name, bundle_s3_key)
@@ -122,7 +122,8 @@ class EC2LaunchManager():
 
     def init_bundle_s3_bucket(self, bundle):
         s3_ctl = self.aim_ctx.get_controller('S3')
-        s3_ctl.init_context(bundle.s3_context_id)
+        bucket_group_name = '-'.join([self.subenv_ctx.netenv_id, self.subenv_id, bundle.app_id, bundle.group_id, bundle.resource_id, self.id])
+        s3_ctl.init_context(self.account_ctx, bundle.s3_context_id, self.subenv_ctx.region, bucket_group_name)
         instance_iam_role_arn_ref = self.subenv_ctx.gen_ref(app_id=bundle.app_id,
                                                   grp_id=bundle.group_id,
                                                   res_id=bundle.resource_id,
@@ -152,11 +153,10 @@ class EC2LaunchManager():
         bucket_config = models.resources.S3Bucket()
         bucket_config.update(bucket_config_dict)
         bucket_config.resolve_ref_obj = self
-        s3_config_ref = '.'.join([self.config_ref, 'applications', bundle.app_id, 'resources', bundle.resource_id, self.id, 'bucket'])
+        bundle.s3_bucket_ref = '.'.join([self.config_ref, 'applications', bundle.app_id, 'resources', bundle.resource_id, self.id, 'bucket'])
         bucket_name_prefix = '-'.join([self.parent.get_aws_name(), bundle.group_id])
         bucket_name_suffix = self.id
         bucket_region = self.subenv_ctx.region
-        s3_config_name = '-'.join([self.subenv_ctx.netenv_id, self.subenv_id, bundle.app_id, bundle.group_id, bundle.resource_id, self.id])
 
         stack_hooks = StackHooks(self.aim_ctx)
         stack_hooks.add('EC2LaunchManager', 'create', 'post',
@@ -164,17 +164,14 @@ class EC2LaunchManager():
         stack_hooks.add('EC2LaunchManager', 'update', 'post',
                         self.stack_hook, self.stack_hook_cache_id, bundle)
 
-        s3_ctl.add_bucket_config(account_ctx=self.account_ctx,
-                                 region=bucket_region,
-                                 app_id=bundle.app_id,
-                                 group_id=bundle.group_id,
-                                 bucket_id=bundle.bucket_id,
-                                 bucket_name_prefix=bucket_name_prefix,
-                                 bucket_name_suffix=bucket_name_suffix,
-                                 config_ref=s3_config_ref,
-                                 config_name=s3_config_name,
-                                 config_dict=bucket_config,
-                                 stack_hooks=stack_hooks)
+        s3_ctl.add_bucket(  bundle.s3_context_id,
+                            region=bucket_region,
+                            bucket_id=bundle.bucket_id,
+                            bucket_name_prefix=bucket_name_prefix,
+                            bucket_name_suffix=bucket_name_suffix,
+                            bucket_ref=bundle.s3_bucket_ref,
+                            bucket_config=bucket_config,
+                            stack_hooks=stack_hooks)
 
 
     def get_s3_bucket_name(self, app_id, grp_id, bucket_id):
@@ -182,8 +179,7 @@ class EC2LaunchManager():
             bundle = self.launch_bundles[bundle_key][0]
             if bundle.app_id == app_id and bundle.group_id == grp_id and bundle.bucket_id == bucket_id:
                 s3_ctl = self.aim_ctx.get_controller('S3')
-                s3_ctl.init_context(bundle.s3_context_id)
-                return s3_ctl.get_bucket_name(bundle.app_id, bundle.group_id, bundle.bucket_id)
+                return s3_ctl.get_bucket_name(bundle.s3_context_id, bundle.s3_bucket_ref)
         return None
 
     def user_data_script(self, app_id, grp_id, resource_id):
@@ -434,20 +430,17 @@ statement:
     def validate(self):
         for s3_context_id in self.launch_bundles.keys():
             s3_ctl = self.aim_ctx.get_controller('S3')
-            s3_ctl.init_context(s3_context_id)
-            s3_ctl.validate()
+            s3_ctl.validate(s3_context_id)
 
     def provision(self):
         for s3_context_id in self.launch_bundles.keys():
             s3_ctl = self.aim_ctx.get_controller('S3')
-            s3_ctl.init_context(s3_context_id)
-            s3_ctl.provision()
+            s3_ctl.provision(s3_context_id)
 
     def delete(self):
         for s3_context_id in self.launch_bundles.keys():
             s3_ctl = self.aim_ctx.get_controller('S3')
-            s3_ctl.init_context(s3_context_id)
-            s3_ctl.delete()
+            s3_ctl.delete(s3_context_id)
 
 class ApplicationStackGroup(StackGroup):
     def __init__(self,
@@ -477,6 +470,10 @@ class ApplicationStackGroup(StackGroup):
         self.cpbd_s3_post_id = self.subenv_ctx.gen_ref(app_id=self.app_id,
                                                        attribute="cpbd.post")
         self.s3_context_id = self.subenv_ctx.gen_ref(app_id=self.app_id)
+        s3_ctl = self.aim_ctx.get_controller('S3')
+        s3_group_name = '-'.join([self.subenv_ctx.netenv_id, self.subenv_id, self.app_id])
+        s3_ctl.init_context(self.account_ctx, self.s3_context_id, self.subenv_ctx.region, s3_group_name)
+
 
         self.ec2_launch_manager = EC2LaunchManager(self.aim_ctx,
                                                    self,
@@ -544,21 +541,17 @@ class ApplicationStackGroup(StackGroup):
             print("ApplicationStackGroup: Init: S3: %s *disabled*" % (res_id))
         else:
             print("ApplicationStackGroup: Init: S3: %s" % (res_id))
-            s3_ctl = self.aim_ctx.get_controller('S3')
-            s3_ctl.init_context(self.s3_context_id)
             s3_config_ref = self.gen_resource_ref(grp_id, res_id)
             # Generate s3 bucket name for application deployment
             bucket_name_prefix = '-'.join([self.get_aws_name(), grp_id])
             #print("Application depoloyment bucket name: %s" % new_name)
-            s3_config_name = '-'.join([self.subenv_ctx.netenv_id, self.subenv_id, self.app_id, grp_id])
-            s3_ctl.add_config(self.account_ctx,
-                                self.aws_region,
-                                bucket_name_prefix,
-                                None,
-                                s3_config_ref,
-                                s3_config_name,
-                                res_id,
-                                res_config)
+            s3_ctl.add_bucket(  self.s3_context_id,
+                                region=self.aws_region,
+                                bucket_id=res_id,
+                                bucket_name_prefix=bucket_name_prefix,
+                                bucket_name_suffix=None,
+                                bucket_ref=s3_config_ref,
+                                bucket_config=res_config)
 
     def init_lbclassic_resource(self, grp_id, res_id, res_config):
         if res_config.enabled == False:
@@ -622,72 +615,72 @@ class ApplicationStackGroup(StackGroup):
             print("ApplicationStackGroup: Init: ASG: %s *disabled*" % (res_id))
         else:
             print("ApplicationStackGroup: Init: ASG: " + res_id)
-            asg_config_ref = self.gen_resource_ref(grp_id, res_id)
-            # Create instance role
-            role_profile_arn = None
-            if res_config.instance_iam_role.enabled == True:
-                iam_ctl = self.aim_ctx.get_controller('IAM')
-                iam_config_ref_prefix = self.subenv_ctx.gen_ref()
-                iam_ctl.init_context(self.account_ctx,
-                                     self.aws_region,
-                                     self.iam_context_id,
-                                     iam_config_ref_prefix)
-                #policy_config = models.iam.IAM('codecommit', res_config.instance_iam_role)
-                #policy_config.add_roles(role_config_dict)
-                # The ID to give this role is: group.resource.instance_iam_role
-                instance_iam_role_ref = self.subenv_ctx.gen_ref(app_id=self.app_id,
-                                                                grp_id=grp_id,
-                                                                res_id=res_id,
-                                                                attribute='instance_iam_role')
-                instance_iam_role_id = self.gen_iam_role_id(grp_id, res_id, 'instance_iam_role')
-                # If no assume policy has been added, force one here since we know its
-                # an EC2 instance using it.
-                role_config = res_config.instance_iam_role
-                # Set defaults if assume role policy was not explicitly configured
-                if not hasattr(role_config, 'assume_role_policy') or role_config.assume_role_policy == None:
-                    policy_dict = { 'effect': 'Allow',
-                                    'service': ['ec2.amazonaws.com'] }
+        asg_config_ref = self.gen_resource_ref(grp_id, res_id)
+        # Create instance role
+        role_profile_arn = None
+        if res_config.instance_iam_role.enabled == True:
+            iam_ctl = self.aim_ctx.get_controller('IAM')
+            iam_config_ref_prefix = self.subenv_ctx.gen_ref()
+            iam_ctl.init_context(self.account_ctx,
+                                    self.aws_region,
+                                    self.iam_context_id,
+                                    iam_config_ref_prefix)
+            #policy_config = models.iam.IAM('codecommit', res_config.instance_iam_role)
+            #policy_config.add_roles(role_config_dict)
+            # The ID to give this role is: group.resource.instance_iam_role
+            instance_iam_role_ref = self.subenv_ctx.gen_ref(app_id=self.app_id,
+                                                            grp_id=grp_id,
+                                                            res_id=res_id,
+                                                            attribute='instance_iam_role')
+            instance_iam_role_id = self.gen_iam_role_id(grp_id, res_id, 'instance_iam_role')
+            # If no assume policy has been added, force one here since we know its
+            # an EC2 instance using it.
+            role_config = res_config.instance_iam_role
+            # Set defaults if assume role policy was not explicitly configured
+            if not hasattr(role_config, 'assume_role_policy') or role_config.assume_role_policy == None:
+                policy_dict = { 'effect': 'Allow',
+                                'service': ['ec2.amazonaws.com'] }
 
-                    role_config.set_assume_role_policy(policy_dict)
-                # Always turn on instance profiles for ASG instances
-                role_config.instance_profile = True
+                role_config.set_assume_role_policy(policy_dict)
+            # Always turn on instance profiles for ASG instances
+            role_config.instance_profile = True
 
-                iam_ctl.add_role(self.iam_context_id,
-                                 self,
-                                 instance_iam_role_id,
-                                 instance_iam_role_ref,
-                                 res_config.instance_iam_role)
-                role_profile_arn = iam_ctl.role_profile_arn(self.iam_context_id, instance_iam_role_id)
-
-            if res_config.monitoring != None:
-                self.ec2_launch_manager.lb_add_cloudwatch_agent(res_config.monitoring,
-                                                                self.app_id,
-                                                                grp_id,
-                                                                res_id,
-                                                                res_config)
-            aws_name = '-'.join([grp_id, res_id])
-            asg_template = aim.cftemplates.ASG(self.aim_ctx,
-                                                self.account_ctx,
-                                                self.subenv_ctx,
-                                                aws_name,
-                                                self.app_id,
-                                                res_id,
-                                                res_config,
-                                                asg_config_ref,
-                                                role_profile_arn,
-                                                self.ec2_launch_manager.user_data_script(self.app_id, grp_id, res_id))
-            asg_stack = Stack(self.aim_ctx,
-                                self.account_ctx,
+            iam_ctl.add_role(self.iam_context_id,
                                 self,
-                                res_config,
-                                asg_template,
-                                aws_region=self.aws_region)
-            self.stack_list.append(asg_stack)
-            self.add_stack_order(asg_stack)
+                                instance_iam_role_id,
+                                instance_iam_role_ref,
+                                res_config.instance_iam_role)
+            role_profile_arn = iam_ctl.role_profile_arn(self.iam_context_id, instance_iam_role_id)
 
-            if res_config.monitoring and len(res_config.monitoring.alarm_sets.values()) > 0:
-                aws_name = '-'.join(['ASG', grp_id, res_id])
-                self.init_alarms(aws_name, asg_config_ref, res_config)
+        if res_config.monitoring != None:
+            self.ec2_launch_manager.lb_add_cloudwatch_agent(res_config.monitoring,
+                                                            self.app_id,
+                                                            grp_id,
+                                                            res_id,
+                                                            res_config)
+        aws_name = '-'.join([grp_id, res_id])
+        asg_template = aim.cftemplates.ASG(self.aim_ctx,
+                                            self.account_ctx,
+                                            self.subenv_ctx,
+                                            aws_name,
+                                            self.app_id,
+                                            res_id,
+                                            res_config,
+                                            asg_config_ref,
+                                            role_profile_arn,
+                                            self.ec2_launch_manager.user_data_script(self.app_id, grp_id, res_id))
+        asg_stack = Stack(self.aim_ctx,
+                            self.account_ctx,
+                            self,
+                            res_config,
+                            asg_template,
+                            aws_region=self.aws_region)
+        self.stack_list.append(asg_stack)
+        self.add_stack_order(asg_stack)
+
+        if res_config.monitoring and len(res_config.monitoring.alarm_sets.values()) > 0:
+            aws_name = '-'.join(['ASG', grp_id, res_id])
+            self.init_alarms(aws_name, asg_config_ref, res_config)
 
     def init_ec2_resource(self, grp_id, res_id, res_config):
         if res_config.enabled == False:
@@ -729,31 +722,28 @@ class ApplicationStackGroup(StackGroup):
             # -----------------
             # S3 Artifacts Bucket:  PRE
             s3_ctl = self.aim_ctx.get_controller('S3')
-            s3_ctl.init_context(self.cpbd_s3_pre_id)
+            s3_group_name = '-'.join([self.subenv_ctx.netenv_id, self.subenv_id, self.app_id, grp_id])
+            s3_ctl.init_context(tools_account_ctx, self.cpbd_s3_pre_id, self.subenv_ctx.region, s3_group_name)
             s3_artifacts_group_id=res_id
             s3_artifacts_bucket_id='artifacts_bucket'
             artifact_bucket_config = res_config.artifacts_bucket
             s3_config_ref = self.gen_resource_ref(grp_id, res_id, s3_artifacts_bucket_id)
             bucket_name_prefix = '-'.join([self.get_aws_name(), grp_id])
-            s3_config_name = '-'.join([self.subenv_ctx.netenv_id, self.subenv_id, self.app_id, grp_id])
             artifact_bucket_config.name = '-'.join([res_id, artifact_bucket_config.name])
-            s3_ctl.add_bucket_config(account_ctx=tools_account_ctx,
-                                    region=self.aws_region,
-                                    app_id=self.app_id,
-                                    group_id=s3_artifacts_group_id,
-                                    bucket_id=s3_artifacts_bucket_id,
-                                    bucket_name_prefix=bucket_name_prefix,
-                                    bucket_name_suffix=None,
-                                    config_ref=s3_config_ref,
-                                    config_name=s3_config_name,
-                                    config_dict=artifact_bucket_config)
+            s3_ctl.add_bucket(self.cpbd_s3_pre_id,
+                                region=self.aws_region,
+                                bucket_id=s3_artifacts_bucket_id,
+                                bucket_name_prefix=bucket_name_prefix,
+                                bucket_name_suffix=None,
+                                bucket_ref=s3_config_ref,
+                                bucket_config=artifact_bucket_config)
 
-            s3_artifacts_bucket_arn=s3_ctl.get_bucket_arn(app_id=self.app_id,
-                                                          group_id=s3_artifacts_group_id,
-                                                          bucket_id=s3_artifacts_bucket_id)
+            s3_artifacts_bucket_arn=s3_ctl.get_bucket_arn(self.cpbd_s3_pre_id,
+                                                            bucket_ref=s3_config_ref)
+
 
             # S3 Artifacts Bucket:  POST
-            s3_ctl.init_context(self.cpbd_s3_post_id)
+            s3_ctl.init_context(tools_account_ctx, self.cpbd_s3_post_id, self.subenv_ctx.region, s3_group_name)
             codebuild_role_ref = self.subenv_ctx.gen_ref(app_id=self.app_id,
                                                          grp_id=grp_id,
                                                          res_id=res_id,
@@ -784,17 +774,13 @@ class ApplicationStackGroup(StackGroup):
 
             artifact_bucket_config.add_policy(cpbd_s3_bucket_policy)
 
-            #artifact_bucket_config['policy'].append(cpbd_s3_bucket_policy)
-            s3_ctl.add_bucket_config(account_ctx=tools_account_ctx,
-                                    region=self.aws_region,
-                                    app_id=self.app_id,
-                                    group_id=s3_artifacts_group_id,
-                                    bucket_id=s3_artifacts_bucket_id,
-                                    bucket_name_prefix=bucket_name_prefix,
-                                    bucket_name_suffix=None,
-                                    config_ref=s3_config_ref,
-                                    config_name=s3_config_name,
-                                    config_dict=artifact_bucket_config)
+            s3_ctl.add_bucket(self.cpbd_s3_post_id,
+                                region=self.aws_region,
+                                bucket_id=s3_artifacts_bucket_id,
+                                bucket_name_prefix=bucket_name_prefix,
+                                bucket_name_suffix=None,
+                                bucket_ref=s3_config_ref,
+                                bucket_config=artifact_bucket_config)
 
 
             # ----------------
@@ -948,7 +934,7 @@ policies:
                                                              grp_id,
                                                              res_id,
                                                              res_config,
-                                                             s3_ctl.get_bucket_name(self.app_id, s3_artifacts_group_id, s3_artifacts_bucket_id),
+                                                             s3_ctl.get_bucket_name(self.cpbd_s3_post_id, s3_config_ref),
                                                              codedeploy_conf_ref)
 
             codedeploy_stack = Stack(self.aim_ctx,
@@ -972,7 +958,7 @@ policies:
                                                                     grp_id,
                                                                     res_id,
                                                                     res_config,
-                                                                    s3_ctl.get_bucket_name(self.app_id, s3_artifacts_group_id, s3_artifacts_bucket_id),
+                                                                    s3_ctl.get_bucket_name(self.cpbd_s3_post_id, s3_config_ref),
                                                                     codedeploy_template.get_tools_delegate_role_arn(),
                                                                     codepipebuild_conf_ref)
 
@@ -1050,19 +1036,15 @@ policies:
 
         # App Group Validation
         s3_ctl = self.aim_ctx.get_controller('S3')
-        s3_ctl.init_context(self.cpbd_s3_pre_id)
-        s3_ctl.validate()
+        s3_ctl.validate(self.cpbd_s3_pre_id)
 
         self.ec2_launch_manager.validate()
 
         super().validate()
 
         # S3 Service Valdiation
-        s3_ctl.init_context(self.s3_context_id)
-        s3_ctl.validate()
-
-        s3_ctl.init_context(self.cpbd_s3_post_id)
-        s3_ctl.validate()
+        s3_ctl.validate(self.s3_context_id)
+        s3_ctl.validate(self.cpbd_s3_post_id)
 
     def provision(self):
         # self.validate()
@@ -1076,8 +1058,7 @@ policies:
         acm_ctl.provision()
 
         s3_ctl = self.aim_ctx.get_controller('S3', self.cpbd_s3_pre_id)
-        s3_ctl.init_context(self.cpbd_s3_pre_id)
-        s3_ctl.provision()
+        s3_ctl.provision(self.cpbd_s3_pre_id)
 
         self.ec2_launch_manager.provision()
 
@@ -1085,25 +1066,20 @@ policies:
         super().provision()
 
         # Provision S3 Services
-        s3_ctl.init_context(self.s3_context_id)
-        s3_ctl.provision()
-
-        s3_ctl.init_context(self.cpbd_s3_post_id)
-        s3_ctl.provision()
+        s3_ctl.provision(self.s3_context_id)
+        s3_ctl.provision(self.cpbd_s3_post_id)
 
     def delete(self):
 
         s3_ctl = self.aim_ctx.get_controller('S3')
-        s3_ctl.init_context(self.cpbd_s3_post_id)
-        s3_ctl.delete()
+        s3_ctl.delete(self.cpbd_s3_post_id)
 
         # S3 Service Valdiation
-        s3_ctl.init_context(self.s3_context_id)
-        s3_ctl.delete()
+        s3_ctl.delete(self.s3_context_id)
 
         # App Group Validation
-        s3_pre_ctl = self.aim_ctx.get_controller('S3', self.cpbd_s3_pre_id)
-        s3_pre_ctl.delete()
+        s3_pre_ctl = self.aim_ctx.get_controller('S3')
+        s3_pre_ctl.delete(self.cpbd_s3_pre_id)
 
         super().delete()
 
@@ -1112,14 +1088,6 @@ policies:
          # IAM
         iam_ctl = self.aim_ctx.get_controller('IAM')
         iam_ctl.delete(self.iam_context_id)
-
-    def get_deployment_artifacts_bucket_value(self, group_id, bucket_id, attribute):
-        s3_ctl = self.aim_ctx.get_controller("S3")
-        s3_ctl.init_context(self.cpbd_s3_pre_id)
-        bucket_name = s3_ctl.get_bucket_name(self.app_id, group_id, bucket_id)
-        if bucket_name == None:
-            raise StackException(AimErrorCode.Unknown)
-        return bucket_name
 
     def get_stack_from_ref(self, ref):
         for stack in self.stack_list: