+++ src/aim/stack_group/grp_s3.py
@@ -9,48 +9,46 @@ class S3StackGroup(StackGroup):
     def __init__(self,
                  aim_ctx,
                  account_ctx,
-                 config,
-                 config_name,
-                 app_id,
-                 group_id,
+                 region,
+                 group_name,
+                 buckets,
                  controller,
                  s3_context_id,
                  stack_hooks=None):
-        aws_name = config_name
+        print("S3 Group Name: " + group_name)
+        aws_name = group_name
         super().__init__(aim_ctx,
                          account_ctx,
-                         config_name,
+                         group_name,
                          aws_name,
                          controller)
 
         # Initialize config with a deepcopy of the project defaults
-        self.config = config
         self.stack_list = []
-        self.app_id = app_id
-        self.group_id = group_id
         self.s3_context_id = s3_context_id
         self.stack_hooks = stack_hooks
+        self.buckets = buckets
+        self.region = region
+
         s3_template = aim.cftemplates.S3(self.aim_ctx,
                                          self.account_ctx,
-                                         config,
-                                         self.app_id,
-                                         self.group_id,
+                                         self.buckets,
                                          self.s3_context_id,
-                                         self.config.config_ref)
-        s3_template.set_template_file_id(self.aim_ctx.md5sum(str_data=self.controller.context_id))
+                                         None)
+        s3_template.set_template_file_id(self.aim_ctx.md5sum(str_data=s3_context_id))
 
         # S3 Delete on Stack Delete hook
         if self.stack_hooks == None:
             self.stack_hooks = StackHooks(self.aim_ctx)
 
         self.stack_hooks.add('S3StackGroup', 'delete', 'post',
-                             self.stack_hook_post_delete, None, config)
+                             self.stack_hook_post_delete, None, self.buckets)
         s3_stack = Stack(self.aim_ctx,
                          self.account_ctx,
                          self,
-                         config,
+                         self.buckets,
                          s3_template,
-                         aws_region=self.config.region,
+                         aws_region=self.region,
                          hooks=self.stack_hooks)
 
         self.stack_list.append(s3_stack)
@@ -59,11 +57,12 @@ class S3StackGroup(StackGroup):
 
     def stack_hook_post_delete(self, hook, hook_arg):
         # Empty the S3 Bucket if enabled
-        s3_config = hook_arg
-        s3_resource = self.account_ctx.get_aws_resource('s3', s3_config.region)
-        for bucket_id in s3_config.get_bucket_ids(self.app_id, self.group_id):
-            deletion_policy = s3_config.get_bucket_deletion_policy(self.app_id, self.group_id, bucket_id)
-            bucket_name = s3_config.get_bucket_name(self.app_id, self.group_id, bucket_id)
+        buckets = hook_arg
+        for bucket_context in buckets:
+            s3_config = bucket_context['config']
+            s3_resource = self.account_ctx.get_aws_resource('s3', self.region)
+            deletion_policy = s3_config.deletion_policy
+            bucket_name = self.controller.get_bucket_name(self.s3_context_id, bucket_context['ref'])
             if deletion_policy == "delete":
                 print("Deleting S3 Bucket: %s" % (bucket_name))
                 bucket = s3_resource.Bucket(bucket_name)
@@ -81,7 +80,6 @@ class S3StackGroup(StackGroup):
         super().validate()
 
     def provision(self):
-        # self.validate()
         super().provision()
 
         return None